openrr_command

Enum RobotCommand

pub enum RobotCommand {

    SendJoints {
        name: String,
        duration: f64,
        use_interpolation: bool,
        joint: Vec<(usize, f64)>,
        max_resolution_for_interpolation: f64,
        min_number_of_points_for_interpolation: i32,
    },
    SendJointsPose {
        name: String,
        pose_name: String,
        duration: f64,
    },
    MoveIk {
        name: String,
        x: Option<f64>,
        y: Option<f64>,
        z: Option<f64>,
        yaw: Option<f64>,
        pitch: Option<f64>,
        roll: Option<f64>,
        duration: f64,
        use_interpolation: bool,
        is_local: bool,
        max_resolution_for_interpolation: f64,
        min_number_of_points_for_interpolation: i32,
    },
    GetState {
        name: String,
    },
    LoadCommands {
        command_file_path: PathBuf,
    },
    List,
    Speak {
        name: String,
        message: Vec<String>,
    },
    ExecuteCommand {
        command: Vec<String>,
    },
    GetNavigationCurrentPose,
    SendNavigationGoal {
        x: f64,
        y: f64,
        yaw: f64,
        frame_id: String,
        timeout_secs: f64,
    },
    CancelNavigationGoal,
    SendBaseVelocity {
        x: f64,
        y: f64,
        theta: f64,
        duration_secs: f64,
    },
    ShellCompletion(ShellType),

}

Variants

SendJoints

Send joint positions.

Fields

name: String

duration: f64

use_interpolation: bool

Interpolate target in cartesian space. If you use this flag, joint values are not used as references but used in forward kinematics.

joint: Vec<(usize, f64)>

Specify joint parameters. Like --joint 0=1.2. In accordance with the sequence in which the “joint names” are defined in the configuration, they are numbered starting at 0.

max_resolution_for_interpolation: f64

min_number_of_points_for_interpolation: i32

SendJointsPose

Send predefined joint positions.

Fields

name: String

pose_name: String

duration: f64

MoveIk

Move with ik

Fields

name: String

x: Option<f64>

y: Option<f64>

z: Option<f64>

yaw: Option<f64>

pitch: Option<f64>

roll: Option<f64>

duration: f64

use_interpolation: bool

Interpolate target in cartesian space.

is_local: bool

max_resolution_for_interpolation: f64

min_number_of_points_for_interpolation: i32

GetState

Get joint positions and end pose if applicable.

Fields

name: String

LoadCommands

Load commands from file and execute them.

Fields

command_file_path: PathBuf

List

List available clients.

Speak

Speak text message.

Fields

name: String

message: Vec<String>

ExecuteCommand

Execute an external command.

Fields

command: Vec<String>

GetNavigationCurrentPose

Get navigation current pose.

SendNavigationGoal

Send navigation goal pose.

Fields

x: f64

y: f64

yaw: f64

frame_id: String

timeout_secs: f64

CancelNavigationGoal

Cancel navigation goal.

SendBaseVelocity

Send base velocity.

Fields

x: f64

y: f64

theta: f64

duration_secs: f64

ShellCompletion(ShellType)

Shell completion